not checking invalid email in auto messages erxes/erxes#1940

+++ src/data/resolvers/mutations/engageUtils.ts
@@ -146,7 +146,7 @@ export const send = async (engageMessage: IEngageMessageDocument) => {
       engageMessageId: engageMessage._id,
     };
 
-    if (customerInfos.length === 0) {
+    if (engageMessage.kind === MESSAGE_KINDS.MANUAL && customerInfos.length === 0) {
       await EngageMessages.deleteOne({ _id: engageMessage._id });
       throw new Error('No customers found who have valid emails');
     }
